+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/dsl/DslNavigationConfig.kt
@@ -329,7 +329,7 @@ internal class DslNavigationConfig(
             parentKey = parentKey?.let { NodeKey(it) },
             stacks = stacks,
             activeStackIndex = config.initialTab,
-            wrapperKey = builder.scopeKey.value,
+            wrapperKey = builder.destinationClass.qualifiedName ?: builder.scopeKey.value,
             tabMetadata = tabMetadata,
             scopeKey = builder.scopeKey
         )